inverse dependency to build NSWindow in carbon render
minor changes
